+++ merlion/models/factory.py
@@ -59,8 +59,8 @@ import_alias = dict(
     MoE_ForecasterEnsemble="merlion.models.ensemble.MoE_forecast:MoE_ForecasterEnsemble",
     # Layers
     SeasonalityLayer="merlion.models.automl.seasonality:SeasonalityLayer",
-    AutoETS="merlion.models.forecast.ets:AutoETS",
-    AutoProphet="merlion.models.forecast.prophet:AutoProphet",
+    AutoETS="merlion.models.automl.autoets:AutoETS",
+    AutoProphet="merlion.models.automl.autoprophet:AutoProphet",
     AutoSarima="merlion.models.automl.autosarima:AutoSarima",
 )
